fix: seperate routers 1. create `router/defaults` dir for seperate 2. modify dev.docker-compose.yml for intergrate dev env at container (linux/amd64)
