+++ app/main.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
 from fastapi.middleware.cors import CORSMiddleware
 
-import router as router
+from router import default_router, settings_router
 from utils import settings
 from utils.logger import logger
 
@@ -14,7 +14,7 @@ app.add_middleware(
     allow_methods=settings.fastapi.allow_methods,
     allow_headers=settings.fastapi.allow_headers,
 )
-app.include_router(router.default_router)
-app.include_router(router.settings_router)
+app.include_router(default_router)
+app.include_router(settings_router)
 app.mount("/data/이지원", StaticFiles(directory="data/이지원"), name="data/이지원")
 logger.debug("Initialized FastAPI App")
