+++ docker/dev.docker-compose.yml
@@ -2,6 +2,7 @@ name: thanosql-ai-chat-dev
 
 services:
   app:
+    platform: linux/amd64
     container_name: thanosql-ai-chat-dev-be
     build:
       context: ..
@@ -25,6 +26,7 @@ services:
       PYTHONUNBUFFERED: 1
 
   redis:
+    platform: linux/amd64
     container_name: thanosql-ai-chat-dev-redis
     build:
       context: redis
